+++ src/ccc/mdebug_importer.h
@@ -13,7 +13,7 @@ namespace ccc::mdebug {
 // Perform all the main analysis passes on the mdebug symbol table and convert
 // it to a set of C++ ASTs.
 Result<SymbolSourceHandle> import_symbol_table(
-	SymbolDatabase& database, const mdebug::SymbolTableReader& symbol_table, u32 parser_flags, DemanglerFunc* demangle);
+	SymbolDatabase& database, const mdebug::SymbolTableReader& symbol_table, u32 parser_flags, const DemanglerFunctions& demangler);
 Result<void> import_files(SymbolDatabase& database, const AnalysisContext& context);
 Result<void> import_file(SymbolDatabase& database, const mdebug::File& input, const AnalysisContext& context);